Challege the method create in the abstract factory

+++ src/main/java/people/Client.java
@@ -2,17 +2,17 @@
 
 package people;
 import item.Item;
-import java.math.BigInteger;
+
 import java.util.List;
 import java.util.ArrayList;
 
 public class Client extends Person{
 
     public double budget;
-    public BigInteger age;
+    public int age;
     public List<Item> itemsPurchased;
 
-    public Client( double budget, BigInteger age) {
+    public Client( double budget, int age) {
 //        this.name = name;
 //        this.id = id;
 //        this.address = address;
